sentiant-platform/python-backend/requirements.txt
flask
vaderSentiment
google-api-python-client
python-dotenv
requests
gunicorn